pyproject.toml
[project]
name = "sds-common"
version = "1.0.4"
description = "A Python library for common functionality used to interact with SDS."
requires-python = "==3.13.*"
dependencies = [
    "cloudevents==1.11.0",
    "google-auth==2.38.0",
    "google-cloud-secret-manager==2.21.1",
    "google-cloud-pubsub==2.28.0",
    "google-cloud-storage==3.0.0",
    "google-cloud-firestore==2.20.0",
    "firebase-admin==6.6.0",
    "requests==2.32.3",
    "pydantic_settings==2.1.0",
    "python-dotenv==1.0.0",
    "functions-framework==3.9.2",
    "pip_audit==2.5.6",
    "setuptools==70.0.0",
    "pytest-order==1.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-xdist>=3.6.1",
]

lint = [
    "ruff>=0.11.7",
]

version-check = [
 "toml>=0.10.2",
 "tomlkit>=0.13.3",
]

[tool.ruff]
line-length = 121
exclude = [ "lib", "__pycache__", ".venv",]

[tool.ruff.lint]
extend-select = [ "E303", "W391", "W292", "E302"]
preview = true

